# server/openapi.yaml
openapi: 3.0.3
info:
  title: BugBoard API
  version: 1.0.0
  description: |
    REST API behind the BugBoard demo app. Every endpoint except
    `POST /api/auth/login`, `GET /api/health` and `POST /api/test/reset`
    requires a bearer token obtained from the login endpoint.
servers:
  - url: http://localhost:4000
    description: Local development
...
paths:
...
  /api/issues/{idOrKey}/attachments:
    parameters:
      - { in: path, name: idOrKey, required: true, schema: { type: string } }
    get:
      tags: [Attachments]
      summary: List attachments on an issue
      responses:
        '200': { description: Attachment list }
    post:
      tags: [Attachments]
      summary: Upload an attachment (max 2 MB)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '201': { description: Uploaded }
        '400': { $ref: '#/components/responses/ValidationError' }
        '413': { description: File exceeds the size limit }
  /api/attachments/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    get:
      tags: [Attachments]
      summary: Download an attachment
      responses:
        '200': { description: File contents }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Attachments]
      summary: Delete an attachment
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  path: { type: string }
                  message: { type: string }
  responses:
    ValidationError:
      description: The request body or query string failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: No such resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
security:
  - bearerAuth: []